// apps/web/src/lib/api/generated/models/getFormFields200DataItem.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * @ding/api
 * Ding API Documentation
 * OpenAPI spec version: 0.0.0
 */

export type GetFormFields200DataItem = {
  id: string
  formId: string
  fieldId: string
  label: string
  /** @nullable */
  intent: string | null
  required: boolean
  sortOrder: number
  createdAt: string
  updatedAt: string
}